add #16 UX向上を目的とした処理を独自クラスに切り分け
